+++ packages/generators/src/loader-generator.ts
@@ -1,6 +1,8 @@
 import path from "path";
 import addonGenerator from "./addon-generator";
 import { toKebabCase } from "./utils/helpers";
+import type { LoaderGeneratorOptions } from "./types";
+import Generator from "yeoman-generator";
 
 /**
  * Formats a string into webpack loader format
@@ -19,16 +21,7 @@ export function makeLoaderName(name: string): string {
   return name;
 }
 
-/**
- * A yeoman generator class for creating a webpack
- * loader project. It adds some starter loader code
- * and runs `webpack-defaults`.
- *
- * @class LoaderGenerator
- * @extends {Generator}
- */
-
-export const LoaderGenerator = addonGenerator(
+export const LoaderGenerator = addonGenerator<LoaderGeneratorOptions>(
   [
     {
       default: "my-loader",
@@ -40,7 +33,9 @@ export const LoaderGenerator = addonGenerator(
     },
   ],
   path.resolve(__dirname, "../loader-template"),
-  (gen): Record<string, unknown> => ({ name: gen.props.name }),
+  (gen): Record<string, unknown> => ({
+    name: (gen.props as Generator.Question).name,
+  }),
 );
 
 export default LoaderGenerator;