+++ packages/info/src/index.ts
@@ -39,10 +39,10 @@ class InfoCommand {
           description: "Adds additional packages to the output",
         },
       ],
-      async (options) => {
+      async (options: { output: string; additionalPackage: string[] }) => {
         let { output } = options;
 
-        const envinfoConfig = {};
+        const envinfoConfig: { [key: string]: boolean } = {};
 
         if (output) {
           // Remove quotes if exist