+++ test/serve/basic/serve-basic.test.js
@@ -548,7 +548,23 @@ describe("basic serve usage", () => {
       },
     );
 
-    expect(normalizeStderr(stderr)).toMatchSnapshot("stderr");
+    // sort logs for CI
+    let normalizedStderr = normalizeStderr(stderr).split("\n");
+    const lastString = normalizedStderr[normalizedStderr.length - 1];
+
+    if (lastString.includes("webpack-dev-middleware")) {
+      [
+        normalizedStderr[normalizedStderr.length - 1],
+        normalizedStderr[normalizedStderr.length - 2],
+      ] = [
+        normalizedStderr[normalizedStderr.length - 2],
+        normalizedStderr[normalizedStderr.length - 1],
+      ];
+    }
+
+    normalizedStderr = normalizedStderr.join("\n");
+
+    expect(normalizedStderr).toMatchSnapshot("stderr");
     expect(stdout).toBeTruthy();
   });
 
@@ -595,7 +611,7 @@ describe("basic serve usage", () => {
     const { stderr, stdout } = await runWatch(__dirname, [
       "serve",
       "--config",
-      "same-ports-dev-serever.config.js",
+      "same-ports-dev-server.config.js",
     ]);
 
     expect(normalizeStderr(stderr)).toMatchSnapshot("stderr");