+++ packages/generators/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.2](https://github.com/webpack/webpack-cli/compare/@webpack-cli/generators@2.4.1...@webpack-cli/generators@2.4.2) (2022-01-24)
+
+**Note:** Version bump only for package @webpack-cli/generators
+
 ## [2.4.1](https://github.com/webpack/webpack-cli/compare/@webpack-cli/generators@2.4.0...@webpack-cli/generators@2.4.1) (2021-10-18)
 
 **Note:** Version bump only for package @webpack-cli/generators